requirements.txt
python-dotenv==1.0.1
torch==2.2.0
torchvision==0.17.0